@@ -6,11 +6,13 @@
   "author": "Łuasz Socha <kontakt@lukasz-socha.pl>",
   "private": true,
   "scripts": {
-    "lint": "eslint --fix --ext .js,.vue ./",
+    "lint": "eslint --fix --ext .js,.vue,.ts ./",
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
     "@quasar/extras": "^1.10.4",
+    "@typescript-eslint/eslint-plugin": "^5.5.0",
+    "@typescript-eslint/parser": "^5.5.0",
     "@vue/compiler-sfc": "^3.0.11",
     "axios": "^0.21.1",
     "chart.js": "^2.9.4",
@@ -23,7 +25,7 @@
   "devDependencies": {
     "@babel/eslint-parser": "^7.0.0",
     "@quasar/app": "^3.1.0",
-    "eslint": "^7.14.0",
+    "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.2",
     "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.19.1",
