@@ -10,47 +10,69 @@
       </SectionHeader>
       <Form
         class="q-mt-md q-mb-lg q-px-md"
-        @submitted="scrollTo"
+        @save="save"
       />
       <Advert />
       <SectionHeader ref="scrollTarget">
         <q-icon name="o_credit_card" />
         Podsumowanie
       </SectionHeader>
-      <Table />
+      <Summary
+        :input="invoiceInputFields"
+      />
       <SectionHeader>
         <q-icon name="o_pie_chart" />
         Wykres
       </SectionHeader>
-      <Statistics />
+      <Statistics
+        :input="invoiceInputFields"
+      />
     </div>
     <Footer />
   </q-page>
 </template>
 
-<script>
-import SectionHeader from 'components/SectionHeader'
-import Form from 'components/invoice/Form'
-import Table from 'components/invoice/Table'
-import Statistics from 'components/invoice/Statistics'
-import Advert from 'components/Advert'
-import Footer from 'components/Footer'
+<script lang="ts">
+import { ref } from 'vue'
+import { useStore } from 'vuex'
+import SectionHeader from 'components/partials/SectionHeader.vue'
+import Form from 'src/components/invoice/Form.vue'
+import Summary from 'components/invoice/Summary.vue'
+import Statistics from 'components/invoice/Statistics.vue'
+import Advert from 'components/Advert.vue'
+import Footer from 'components/Footer.vue'
 import helpers from 'src/logic/helpers'
+import { InvoiceInputFields } from 'src/components/invoice/interfaces/InvoiceInputFields'
+import constants from 'src/logic/constants'
 
 export default {
-  created () {
-    this.$store.commit('app/SET_MODULE_TITLE', 'Faktura VAT')
-    this.$store.commit('invoice/clearAllData')
-  },
-  methods: {
-    scrollTo () {
-      helpers.scrollToElement(this.$refs.scrollTarget.$el)
-    },
+  setup() {
+    const store = useStore()
+    store.commit('app/SET_MODULE_TITLE', 'Faktura VAT')
+
+    const invoiceInputFields = ref(<InvoiceInputFields>{
+      amount: 0,
+      amountType: constants.AMOUNT_TYPES.NET,
+      taxRate: 0,
+    })
+
+    const scrollTarget = ref(null) as any
+
+    const save = (input:InvoiceInputFields) => {
+      invoiceInputFields.value = input
+      helpers.scrollToElement(scrollTarget?.value?.$el)
+    }
+
+    return{
+      invoiceInputFields,
+      scrollTarget,
+      save,
+    }
   },
   components: {
     SectionHeader,
     Form,
-    Table,
+    Summary,
     Statistics,
     Advert,
     Footer,
