@@ -1,53 +1,37 @@
+const { resolve } = require('path')
+
 module.exports = {
-  // https://eslint.org/docs/user-guide/configuring#configuration-cascading-and-hierarchy
-  // This option interrupts the configuration hierarchy at this file
-  // Remove this if you have an higher level ESLint config file (it usually happens into a monorepos)
   root: true,
 
   parserOptions: {
-    parser: '@babel/eslint-parser',
-    ecmaVersion: 2018, // Allows for the parsing of modern ECMAScript features
+    parser: '@typescript-eslint/parser',
     sourceType: 'module', // Allows for the use of imports
   },
 
   env: {
     browser: true,
+    node: true
   },
 
-  // Rules order is important, please avoid shuffling them
+  // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
+  // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
   extends: [
-    // Base ESLint recommended rules
-    // 'eslint:recommended',
-
-    // Uncomment any of the lines below to choose desired strictness,
-    // but leave only one uncommented!
-    // See https://eslint.vuejs.org/rules/#available-rules
-    'plugin:vue/vue3-essential', // Priority A: Essential (Error Prevention)
-    // 'plugin:vue/vue3-strongly-recommended', // Priority B: Strongly Recommended (Improving Readability)
-    // 'plugin:vue/vue3-recommended', // Priority C: Recommended (Minimizing Arbitrary Choices and Cognitive Overhead)
-
-    'standard',
-
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:vue/essential'
   ],
 
-  plugins: [
-    // https://eslint.vuejs.org/user-guide/#why-doesn-t-it-work-on-vue-file
-    // required to lint *.vue files
-    'vue',
-
-  ],
+  // required to lint *.vue files
+  plugins: ['vue', '@typescript-eslint'],
 
   globals: {
     ga: true, // Google Analytics
     cordova: true,
     __statics: true,
-    process: true,
-    Capacitor: true,
-    chrome: true,
+    process: true
   },
 
   // add your custom rules here
-
   rules: {
     'max-len': 0,
     // TODO: Remove when https://github.com/babel/babel-eslint/issues/530 is fixed
@@ -133,5 +117,17 @@ module.exports = {
         ],
       },
     ],
-  },
-}
+    'prefer-promise-reject-errors': 'off',
+
+    // allow console.log during development only
+    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    // allow debugger during development only
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+
+    // TypeScript
+    "no-unsafe-any": 'off',
+    'quotes': ['warn', 'single'],
+    '@typescript-eslint/explicit-function-return-type': 'off',
+    '@typescript-eslint/no-explicit-any': "off",
+  }
+};
