@@ -1,7 +1,7 @@
 import ga from '../logic/analytics.js'
 
 export default ({ router }) => {
-  router.afterEach((to, from) => {
+  router.afterEach((to) => {
     const path = 'app' + to.path
     ga.logPage(path)
   })
