@@ -19,12 +19,12 @@
         <div class="q-mt-sm block">
           <q-radio
             v-model="amountType"
-            val="net"
+            :val="constants.AMOUNT_TYPES.NET"
             label="netto"
           />
           <q-radio
             v-model="amountType"
-            val="gross"
+            :val="constants.AMOUNT_TYPES.GROSS"
             label="brutto"
           />
         </div>
@@ -32,7 +32,7 @@
       <div class="col-12 col-md-6 q-pl-md-sm">
         <q-select
           v-model="taxRate"
-          :options="constants.VAT_VALUES"
+          :options="vatTaxRates"
           label="Stawka podatku VAT*"
           color="brand"
           required
@@ -54,40 +54,60 @@
   </q-form>
 </template>
 
-<script>
+<script lang="ts">
+import {computed, ref} from 'vue'
 import constants from 'src/logic/constants'
+import { InvoiceInputFields } from './interfaces/InvoiceInputFields'
+import {AmountType} from 'src/types/AmountType'
+import {VatTaxRate} from 'src/types/VatTaxRate'
 
 export default {
-  emits: ['submitted'],
-  setup () {
-    return { constants }
-  },
-  data () {
-    return {
-      amount: null,
-      amountType: 'net',
-      taxRate: null,
-    }
-  },
-  created () {
-    this.taxRate = this.constants.DEFAULT_VAT_VALUE
-  },
-  computed: {
-    isDisabledButton () {
-      if (!this.amount || this.taxRate === null) {
-        return true
-      }
-      return false
-    },
-  },
-  methods: {
-    save () {
-      this.$store.commit('invoice/setAmount', Number(this.amount))
-      this.$store.commit('invoice/setAmountType', this.amountType)
-      this.$store.commit('invoice/setTaxRate', Number(this.taxRate.value) / 100)
+  setup(props: any, context: any) {
+    const vatTaxRates = [
+      {
+        label: '0%',
+        value: 0,
+      },
+      {
+        label: '5%',
+        value: 0.05,
+      },
+      {
+        label: '8%',
+        value: 0.08,
+      },
+      {
+        label: '23%',
+        value: 0.23,
+      },
+    ]
 
-      this.$emit('submitted')
-    },
+    const amount = ref(null)
+    const amountType = ref(constants.AMOUNT_TYPES.NET)
+    const taxRate = ref(vatTaxRates[vatTaxRates.length - 1])
+
+    const isDisabledButton = computed(() => {
+      return !amount.value || taxRate.value === null
+    })
+
+    const save = () => {
+      const input: InvoiceInputFields = {
+        amount: Number(amount.value),
+        amountType: amountType.value as AmountType,
+        taxRate: taxRate.value.value as VatTaxRate,
+      }
+      context.emit('save', input)
+    }
+
+    return {
+      constants,
+      save,
+      vatTaxRates,
+      amount,
+      amountType,
+      taxRate,
+      isDisabledButton,
+    }
   },
 }
 </script>
