@@ -176,29 +176,7 @@ const from2022 = {
 export default {
   APP: {
     NAME: 'Kalkulator finansowy',
-    VERSION: '4.2.0',
-  },
-  VAT_VALUES: [
-    {
-      label: '0%',
-      value: 0,
-    },
-    {
-      label: '5%',
-      value: 5,
-    },
-    {
-      label: '8%',
-      value: 8,
-    },
-    {
-      label: '23%',
-      value: 23,
-    },
-  ],
-  DEFAULT_VAT_VALUE: {
-    label: '23%',
-    value: 23,
+    VERSION: '4.3.0',
   },
   AMOUNT_TYPES: {
     NET: 'net',
@@ -330,140 +308,6 @@ export default {
   CASH_REGISTER_LIMIT: 20000,
   VAT_LIMIT: 200000,
   MINIMUM_SALARY: 2800,
-  MENU: {
-    BUSINESS: [
-      {
-        title: 'Samozatrudnienie',
-        caption: 'Moduł oblicza składowe wynagrodzenia przy samozatrudnieniu',
-        className: 'c-selfEmployment',
-        link: '/samozatrudnienie',
-      },
-      {
-        title: 'Działalność niezarejestrowana',
-        caption: 'Moduł oblicza składowe wynagrodzenia przy działalności niezarejestrowanej',
-        className: 'c-unregisteredCompany',
-        link: '/dzialalnosc-niezarejestrowana',
-      },
-      {
-        title: 'Faktura VAT',
-        caption: 'Moduł przelicza kwoty z netto na brutto i z brutto na netto',
-        className: 'c-invoice',
-        link: '/faktura-vat',
-      },
-      {
-        title: 'Limit zwolnienia z ewidencjonowania w kasie fiskalnej',
-        caption: 'Moduł oblicza kwotę obrotu, której przekroczenie powoduje utratę prawa do zwolnienia z ewidencjonowania',
-        className: 'c-cashRegisterLimit',
-        link: '/limit-obrotu-dla-kasy-fiskalnej',
-      },
-      {
-        title: 'Limit zwolnienia z VAT',
-        caption: 'Moduł oblicza kwotę przychodu, którego przekroczenie powoduje utratę prawa do zwolnienia z VAT',
-        className: 'c-vatLimit',
-        link: '/limit-zwolnienia-z-vat',
-      },
-      {
-        title: 'Składki ZUS za część miesiąca',
-        caption: 'Moduł oblicza składki ZUS dla przedsiębiorców za część miesiąca',
-        className: 'c-partialZusContributions',
-        link: '/skladki-zus-za-czesc-miesiaca',
-      },
-    ],
-    WORK: [
-      {
-        title: 'Umowa o pracę',
-        caption: 'Moduł oblicza składowe wynagrodzenia przy umowie o pracę',
-        className: 'c-contractOfEmployment',
-        link: '/umowa-o-prace',
-      },
-      {
-        title: 'Umowa zlecenie',
-        caption: 'Moduł oblicza składowe wynagrodzenia przy umowie zlecenie',
-        className: 'c-contractOfMandate',
-        link: '/umowa-zlecenie',
-      },
-      {
-        title: 'Umowa o dzieło',
-        caption: 'Moduł oblicza składowe wynagrodzenia przy umowie o dzieło',
-        className: 'c-contractWork',
-        link: '/umowa-o-dzielo',
-      },
-      {
-        title: 'Rozliczenie z małżonkiem',
-        caption: 'Moduł umożliwia obliczenie podatku dochodowego przy wspólnym rozliczeniu małżonków',
-        className: 'c-work',
-        link: '/rozliczenie-z-malzonkiem',
-      },
-      {
-        title: 'Zasiłek chorobowy',
-        caption: 'Moduł oblicza wysokość zasiłku chorobowego',
-        className: 'c-sickPay',
-        link: '/zasilek-chorobowy',
-      },
-    ],
-    PERCENTAGE: [
-      {
-        title: 'Lokata',
-        caption: 'Moduł oblicza zysk z lokaty',
-        className: 'c-investment',
-        link: '/lokata',
-      },
-      {
-        title: 'Odsetki',
-        caption: 'Moduł oblicza odsetki za każdy dzień',
-        className: 'c-interest',
-        link: '/odsetki',
-      },
-    ],
-    CURRENCIES: [
-      {
-        title: 'Kursy walut',
-        caption: 'Moduł pokazuje kursy walut',
-        className: 'c-exchangeRates',
-        link: '/kursy-walut',
-      },
-      {
-        title: 'Przelicznik walut',
-        caption: 'Moduł służy do przeliczania jednej waluty na drugą',
-        className: 'c-currencyConverter',
-        link: '/przelicznik-walut',
-      },
-    ],
-    APP: [
-      {
-        title: 'Historia zmian',
-        caption: 'Pokazuje historię zmian w aplikacji',
-        className: 'c-changeLogs',
-        link: '/historia-zmian',
-      },
-      {
-        title: 'Formularz kontaktowy',
-        caption: 'Formularz do zgłaszania błędów lub pomysłów na nowe funkcje',
-        className: 'c-contact',
-        link: '/kontakt',
-      },
-      {
-        title: 'O aplikacji',
-        caption: '',
-        className: 'c-aboutApp',
-        link: 'https://kalkulatorfinansowy.app',
-      },
-      {
-        title: 'Polub na Facebooku',
-        caption: '@kalkulator.finansowy',
-        className: 'c-aboutApp',
-        link: 'https://www.facebook.com/kalkulator.finansowy',
-      },
-    ],
-    INFORMATOR: [
-      {
-        title: 'Inflacja',
-        caption: 'Moduł pokazuje wykres z inflacją w Polsce',
-        className: 'c-informator',
-        link: '/inflacja',
-      },
-    ],
-  },
   LOCALE_DATE: {
     /* starting with Sunday */
     days: 'Niedziela_Poniedziałek_Wtorek_Środa_Czwartek_Piątek_Sobota'.split('_'),
