@@ -111,9 +111,9 @@
 import { ref } from 'vue'
 import helpers from 'src/logic/helpers'
 import constants from 'src/logic/constants'
-import ChooseYear from 'src/components/ChooseYear'
+import ChooseYear from 'src/components/partials/ChooseYear'
 import FormsOfAccounting from 'src/components/partials/FormsOfAccounting'
-import SectionHeader from 'components/SectionHeader'
+import SectionHeader from 'components/partials/SectionHeader'
 import Footer from 'components/Footer'
 import ContractOfEmploymentForm from 'src/components/contractOfEmployment/Form'
 import SelfEmploymentForm from 'src/components/selfEmployment/Form'
