@@ -36,7 +36,7 @@
 
 <script>
 import logs from 'components/changeLogs/logs'
-import SectionHeader from 'components/SectionHeader'
+import SectionHeader from 'components/partials/SectionHeader'
 import ChangeLog from 'components/changeLogs/ChangeLog'
 import Advert from 'components/Advert'
 import Footer from 'components/Footer'
