@@ -86,9 +86,9 @@
 <script>
 import { mapGetters } from 'vuex'
 import helpers from 'src/logic/helpers'
-import SectionHeader from 'components/SectionHeader'
+import SectionHeader from 'components/partials/SectionHeader'
 import Advert from 'components/Advert'
-import ChooseYear from 'components/ChooseYear'
+import ChooseYear from 'components/partials/ChooseYear'
 import Form from 'components/contractOfEmployment/Form'
 import SummarySalaryTable from 'components/contractOfEmployment/SummarySalaryTable'
 import EmployeeTable from 'components/contractOfEmployment/EmployeeTable'
