@@ -86,9 +86,9 @@
 <script>
 import { ref } from 'vue'
 import { mapGetters } from 'vuex'
-import SectionHeader from 'components/SectionHeader'
+import SectionHeader from 'components/partials/SectionHeader'
 import Advert from 'components/Advert'
-import ChooseYear from 'components/ChooseYear'
+import ChooseYear from 'components/partials/ChooseYear'
 import Form from 'components/contractOfMandate/Form'
 import SummarySalaryTable from 'components/contractOfMandate/SummarySalaryTable'
 import EmployeeTable from 'components/contractOfMandate/EmployeeTable'
