@@ -37,8 +37,8 @@
 
 <script>
 import { ref } from 'vue'
-import ChooseYear from 'src/components/ChooseYear'
-import SectionHeader from 'components/SectionHeader'
+import ChooseYear from 'src/components/partials/ChooseYear'
+import SectionHeader from 'components/partials/SectionHeader'
 import Advert from 'components/Advert'
 import Form from 'components/contractWork/Form'
 import Table from 'components/contractWork/Table'
