@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import SectionHeader from 'components/SectionHeader'
+import SectionHeader from 'components/partials/SectionHeader'
 import Form from 'components/investment/Form'
 import Table from 'components/investment/Table'
 import Advert from 'components/Advert'
