@@ -4,10 +4,10 @@ export default {
   loadLatestExchangeRates () {
     return axios.get('https://api.nbp.pl/api/exchangerates/tables/a')
   },
-  loadExchangeRatesFromDate ({ commit }, { date }) {
+  loadExchangeRatesFromDate ({  }, { date }) {
     return axios.get(`https://api.nbp.pl/api/exchangerates/tables/a/${date}`)
   },
-  loadExchangeRateCurrency ({ commit }, { code, startDate, endDate }) {
+  loadExchangeRateCurrency ({  }, { code, startDate, endDate }) {
     return axios.get(`https://api.nbp.pl/api/exchangerates/rates/a/${code}/${startDate}/${endDate}`)
   },
 }
